refactor pkgs
