+++ gradle/libs.versions.toml
@@ -10,7 +10,7 @@ activityCompose = "1.10.1"
 composeBom = "2025.04.00"
 coil = "3.0.0-alpha06"
 ktor = "2.3.11"
-navigation-compose= "2.8.0-alpha02"
+navigation-compose = "2.8.0-alpha02"
 androidx-lifecycle = "2.8.4"
 kotlinx-serialization = "1.6.3"
 media3CommonKtx = "1.5.1"
